Refactor: Improve File/Function Correlation in /v2 #25
